// frontend/src/app/models/class.ts
export class Class {
    descr: string = "";
    datetime: string = "";
    student: string = "";
    teacher: string = "";
    status: string = "";
    subject: string = "";
    double: number = 0;
    explanation: string = "";
    comment: string = "";
    review: number = 0;
    teacherComment: string = "";
}